FIX: Refactor the repo
