--- mantra/augmentations/__init__.py
@@ -1,4 +1,0 @@
-"""Augmentations module for generating new triangulations."""
-
-from .balancing import balance_dataset as balance_dataset
-from .triangulation import Triangulation as Triangulation
