+++ mantra/datasets/__init__.py
@@ -1,5 +1,5 @@
 from .cy import CY
 from .mantra import ManifoldTriangulations
-from .mantra_divided import MANTRADivided
+from .mantra_dataset import MantraDataset
 
-__all__ = ["ManifoldTriangulations", "CY", "MANTRADivided"]
+__all__ = ["ManifoldTriangulations", "CY", "MantraDataset"]
