+++ mantra/representations/graph/hasse_diagram.py
