+++ mantra/representations/graph/one_skeleton.py
