+++ mantra/representations/simplicial/simplicial_connectivity.py
