--- mantra/tasks/task_types.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class TaskType(Enum):
-    NAME = "name"
-    ORIENTABILITY = "orientability"
-    BETTI_NUMBERS = "betti_numbers"
-    HOMEOMORPHIC_DIST = "homeomorphic_dist"
