+++ mantra/utils/__init__.py
@@ -1,0 +1,5 @@
+from .triangulation import Triangulation
+
+__all__ = [
+    "Triangulation"
+]
