+++ mantra/utils/balancing.py
@@ -7,14 +7,14 @@ import sys
 from collections import defaultdict
 from typing import Dict, List, Tuple
 
-from mantra.augmentations.constants import (
+from mantra.manifold_types import Manifold2Type
+from mantra.utils.constants import (
     BETTI_NUMBERS,
     CROSSCAP_GLUE_MAP,
     TORUS_GLUE_MAP,
 )
-from mantra.augmentations.triangulation import Triangulation
-from mantra.manifold_types import Manifold2Type
 from mantra.utils.deduplication import find_duplicates
+from mantra.utils.triangulation import Triangulation
 
 GLUE_ADDS_N_VERTICES = {"torus": 3, "crosscap": 1}
