+++ mantra/utils/triangulation.py
@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 from collections import defaultdict
 from itertools import combinations, permutations
 
-from mantra.augmentations.constants import (
+from mantra.utils.constants import (
     RP2_TRIANGULATION_MINUS_FACE,
     TORUS_TRIANGULATION_MINUS_FACE,
 )
